Move SubskeletonDescriptionInterpolationPlayer in CDK.Description
