+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Animation/Loader/SkeletonAnimationNodeLoader.cs
@@ -19,7 +19,7 @@ using inetum.unityUtils;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-
+using umi3d.cdk.userCapture.description;
 using umi3d.common;
 using umi3d.common.userCapture;
 using umi3d.common.userCapture.animation;
