+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/UserCapture/Runtime/Skeleton/Pose/Utils/PoseClipAggregator.cs
@@ -19,7 +19,7 @@ using inetum.unityUtils;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-
+using umi3d.cdk.userCapture.description;
 using umi3d.common.userCapture.description;
 
 namespace umi3d.cdk.userCapture.pose
